Extract all matrix op to a standalone file
