+++ src/Ext/Aircraft/Hooks.cpp
@@ -298,31 +298,6 @@ DEFINE_HOOK(0x44402E, BuildingClass_ExitObject_PoseDir2, 0x5)
 
 #pragma endregion
 
-DEFINE_HOOK(0x4CF68D, FlyLocomotionClass_DrawMatrix_OnAirport, 0x5)
-{
-	GET(ILocomotion*, iloco, ESI);
-	__assume(iloco != nullptr);
-	auto loco = static_cast<FlyLocomotionClass*>(iloco);
-	auto pThis = static_cast<AircraftClass*>(loco->LinkedTo);
-	if (pThis->GetHeight() <= 0)
-	{
-		REF_STACK(Matrix3D, mat, STACK_OFFSET(0x38, -0x30));
-		auto slope_idx = MapClass::Instance->GetCellAt(pThis->Location)->SlopeIndex;
-		mat = Matrix3D::VoxelRampMatrix[slope_idx] * mat;
-		float ars = pThis->AngleRotatedSideways;
-		float arf = pThis->AngleRotatedForwards;
-		if (std::abs(ars) > 0.005 || std::abs(arf) > 0.005)
-		{
-			mat.TranslateZ(float(std::abs(Math::sin(ars)) * pThis->Type->VoxelScaleX
-				+ std::abs(Math::sin(arf)) * pThis->Type->VoxelScaleY));
-			R->ECX(pThis);
-			return 0x4CF6AD;
-		}
-	}
-
-	return 0x4CF6A0;
-}
-
 DEFINE_HOOK(0x415EEE, AircraftClass_Fire_KickOutPassengers, 0x6)
 {
 	enum { SkipKickOutPassengers = 0x415F08 };
